Add rootName tests
